pyproject.toml
[project]
name = "vilomem"
version = "0.1.0"
description = "Vision-Language Memory Agent for evaluating multimodal models on math and visual reasoning tasks"
authors = [
    { name = "Weihao Bo", email = "weihao.bo@example.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
urls = { Homepage = "https://github.com/weihao-bo/ViLoMem", Repository = "https://github.com/weihao-bo/ViLoMem" }
dependencies = [
    "langgraph>=0.6.0,<0.7.0",
    "langchain-openai>=0.1.22",
    "langchain>=0.2.14",
    "python-dotenv>=1.0.1",
    "langchain-qwq>=0.2.1",
    "math-verify>=0.8.0",
    "langchain-community>=0.3.31",
    "dashscope>=1.24.9",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "open-clip-torch>=2.31.0",
    "accelerate>=1.1.0",
    "qwen-vl-utils>=0.0.14",
    "transformers>=4.44.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "datasets>=4.4.0",
    "scikit-learn>=1.7.2",
    "tqdm>=4.66.0",
    "langchain-siliconflow>=0.1.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest>=8.3.5", "pytest-asyncio>=0.23.0"]
vlmeval = ["ms-vlmeval"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["common", "vl_agent", "vl_agent_baseline", "tools"]

[tool.setuptools.package-dir]
"common" = "src/common"
"vl_agent" = "src/vl_agent"
"vl_agent_baseline" = "src/vl_agent_baseline"
"tools" = "tools"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
